Don't ignore essential Eclipse project files.
The .project and important .classpath files are a part of an Android Eclipse
project, and their changes should be stored in the repository. Although there
may be a case for ignoring these files in very rare situations, they don't
justify belonging in a generic template.

The .project file is generally recreatable with a new import without
difficulty, but it still may have important configuration such as extra build
steps. More importantly, .classpath is absolutely essential for a project to
build correctly and cannot be automatically generated unless there is nothing
extra on the classpath.

Reverts commit 7f6a73723.