Enable Ignore of NuGet Packages Folder By Default
Submitting this for consideration.

The first thing that I do on every new project is to go in and uncomment the packages folder. With the prevalence of NuGet and prominence of its use in the Visual Studio environment, along with how well package restore *just works* now, I believe this should be the default.

While I understand that package restore isn't on by default, I would argue that the types of developers using NuGet _and_ a distributed SCM are the types of developers that would omit the binaries from source control.