Updated Zend Framework gitignore as per docs
```
data/: This directory provides a place to store application data that is volatile and possibly temporary. The disturbance of data in this directory might cause the application to fail. Also, the information in this directory may or may not be committed to a subversion repository. Examples of things in this directory are session files, cache files, sqlite databases, logs and indexes.
```

As per docs...

http://framework.zend.com/manual/2.3/en/ref/project.structure.html